Add sync committee implementation, addressed #1 #2

+++ setup.sh
@@ -1,5 +1,17 @@
 #!/bin/bash
 
+file=/etc/systemd/system/eth2_block_monitoring.service
+if test -f "$file"; then
+    while true; do
+        read -p "Service file already exists. Would you like to replace it? [Y/N] " yn
+        case $yn in
+            [Yy]* ) break;;
+            [Nn]* ) exit;;
+            * ) echo "Please enter Y/N.";;
+        esac
+    done
+fi
+
 tee /etc/systemd/system/eth2_block_monitoring.service > /dev/null <<EOF  
 [Unit]
 Description     = ETH2 Block Monitoring Service
@@ -19,4 +31,3 @@ EOF
 systemctl daemon-reload
 systemctl enable eth2_block_monitoring
 systemctl start eth2_block_monitoring
-
